pyproject.toml
[project]
name = "leadgenerator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "langchain>=1.1.3",
    "langchain-openai>=1.1.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.52.1",
    "tiktoken>=0.12.0",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "flake8>=7.3.0",
    "mypy>=1.19.0",
    "pytest>=9.0.2",
]


